+++ src/cosmosdb-preview/azext_cosmosdb_preview/_params.py
@@ -865,17 +865,17 @@ def load_arguments(self, _):
 
     with self.argument_context('cosmosdb softdeleted-account show') as c:
         c.argument('location', options_list=['--location', '-l'], help="Location of the soft-deleted account.", required=True)
-        c.argument('account_name', options_list=['--account-name', '-n'], help="Name of the soft-deleted Cosmos DB account.", required=True)
+        c.argument('account_name', options_list=['--name', '-n'], help="Name of the soft-deleted Cosmos DB account.", required=True)
         c.argument('resource_group', options_list=['--resource-group', '-g'], help="Name of the resource group.", required=True)
 
     with self.argument_context('cosmosdb softdeleted-account delete') as c:
         c.argument('location', options_list=['--location', '-l'], help="Location of the soft-deleted account.", required=True)
-        c.argument('account_name', options_list=['--account-name', '-n'], help="Name of the soft-deleted Cosmos DB account to purge.", required=True)
+        c.argument('account_name', options_list=['--name', '-n'], help="Name of the soft-deleted Cosmos DB account to purge.", required=True)
         c.argument('resource_group', options_list=['--resource-group', '-g'], help="Name of the resource group.", required=True)
 
     with self.argument_context('cosmosdb softdeleted-account recover') as c:
         c.argument('location', options_list=['--location', '-l'], help="Location of the soft-deleted account.", required=True)
-        c.argument('account_name', options_list=['--account-name', '-n'], help="Name of the soft-deleted Cosmos DB account to recover.", required=True)
+        c.argument('account_name', options_list=['--name', '-n'], help="Name of the soft-deleted Cosmos DB account to recover.", required=True)
         c.argument('resource_group', options_list=['--resource-group', '-g'], help="Name of the resource group.", required=True)
 
     # Soft-deleted Database
@@ -903,27 +903,27 @@ def load_arguments(self, _):
         c.argument('resource_group', options_list=['--resource-group', '-g'], help="Name of the resource group.", required=True)
 
     # Soft-deleted Collection
-    with self.argument_context('cosmosdb sql softdeleted-collection list') as c:
+    with self.argument_context('cosmosdb sql softdeleted-container list') as c:
         c.argument('location', options_list=['--location', '-l'], help="Location of the account.", required=True)
         c.argument('account_name', options_list=['--account-name', '-a'], help="Name of the Cosmos DB account.", required=True)
         c.argument('database_name', options_list=['--database-name', '-d'], help="Name of the database.", required=True)
         c.argument('resource_group', options_list=['--resource-group', '-g'], help="Name of the resource group.", required=True)
 
-    with self.argument_context('cosmosdb sql softdeleted-collection show') as c:
+    with self.argument_context('cosmosdb sql softdeleted-container show') as c:
         c.argument('location', options_list=['--location', '-l'], help="Location of the account.", required=True)
         c.argument('account_name', options_list=['--account-name', '-a'], help="Name of the Cosmos DB account.", required=True)
         c.argument('database_name', options_list=['--database-name', '-d'], help="Name of the database.", required=True)
         c.argument('container_name', options_list=['--name', '-n'], help="Name of the soft-deleted container.", required=True)
         c.argument('resource_group', options_list=['--resource-group', '-g'], help="Name of the resource group.", required=True)
 
-    with self.argument_context('cosmosdb sql softdeleted-collection delete') as c:
+    with self.argument_context('cosmosdb sql softdeleted-container delete') as c:
         c.argument('location', options_list=['--location', '-l'], help="Location of the account.", required=True)
         c.argument('account_name', options_list=['--account-name', '-a'], help="Name of the Cosmos DB account.", required=True)
         c.argument('database_name', options_list=['--database-name', '-d'], help="Name of the database.", required=True)
         c.argument('container_name', options_list=['--name', '-n'], help="Name of the soft-deleted container to purge.", required=True)
         c.argument('resource_group', options_list=['--resource-group', '-g'], help="Name of the resource group.", required=True)
 
-    with self.argument_context('cosmosdb sql softdeleted-collection recover') as c:
+    with self.argument_context('cosmosdb sql softdeleted-container recover') as c:
         c.argument('location', options_list=['--location', '-l'], help="Location of the account.", required=True)
         c.argument('account_name', options_list=['--account-name', '-a'], help="Name of the Cosmos DB account.", required=True)
         c.argument('database_name', options_list=['--database-name', '-d'], help="Name of the database.", required=True)